Ignore NPE in MapLoaderTest if it is caused by a known issue
Ignoring NPE with message `Key loaded by a MapLoader cannot be null.` in some test cases where it is known that an enforced exception may be rethrown by subsequent Map operations due to a known issue with MapLoader. When it happens, it eliminates the assertions from the test case. Since this issue occurs rarely, ignoring the NPE in this case is better than ignoring the test. This way the test cases still cover the related MapLoader functionalities in most of the executions.

Related 11244, hazelcast/hazelcast-enterprise/issues/2105