Fixed possible race condition in Near Cache serialization count tests
The Near Cache serialization count tests fail, when the invalidations
from the initial population are sent and executed (because we'll have
to re-populate the Near Cache in that case, which causes additional
(de)serializations).

We can fix this by increasing the batch size and frequency seconds to
very large values, so no invalidations are executed during the test.

In addition we can use the smallInstanceConfig() to reduce the resource
usage of the tests.