Remove assumption that IO thread can have only 1 dead pipeline
The test assumes the IO balancer will balance both the "live" and the
"dead" pipelines. This might not be the case and an IO thread can have
a high number of dead pipelines as they do not count much against the
load on the IO thread.
The fix changes the thread assertions to allow only one "live" pipeline
and more than one "dead" pipeline.

Fixes: https://github.com/hazelcast/hazelcast/issues/7662