Lower heartbeat interval/timeout for quorum tests
During cluster split done in quorum tests we may encounter the
following situation:

1. Communication between split sub-clusters is blocked by blacklisting
the sub-cluster members and closing the active connections between them.

2. To force the formation of the sub-clusters, all the members of one
sub-cluster are marked as suspicious in another one and wise versa.

3. But there might be "phantom" heartbeat operations queued for the
execution, but not yet executed on the members, such heartbeats can
remove the suspicion from the member. This delays the sub-cluster
formation until a heartbeat timeout will be triggered. Which is 60
seconds by default, but the tests expect the sub-cluster formation to be
finished under 30 seconds.

This change lowers heartbeat interval/timeout and increases the
expected cluster formation time to avoid sub-cluster formation timeouts.

Fixes: https://github.com/hazelcast/hazelcast/issues/13097