Fixes cache config propagation on a mixed 3.9 - 3.10 cluster
While doing a rolling upgrade with a mixed 3.9 - 3.10 cluster running
on 3.9 cluster version, cache configs must be converted to CacheConfig
before dispatched to 3.9 members. Otherwise, a PreJoinCacheConfig could
be stored on a 3.9 member's cache configs registry
(AbstractCacheService#configs) and from there it may leak to a
client getting the cache config causing deserialization errors on the
client side. Compatibility test is added in enterprise-side PR.