Avoid full scans induced by paging predicate
If the inner predicate of the paging predicate returns an empty result
set, the paging predicate returns null as the indexed query result, this
interpreted as an absence of the indexed result by the query engine and
the execution falls back to the full scan. This change fixes this
misbehavior.

(cherry-picked from b24588a1b32e6ce449629eb7b337b9af36be85d6)

Fixes: https://github.com/hazelcast/hazelcast/issues/13159