Fixes tmp deadlock on NioOutboundPipeline.close
If the close method gets called from the NIO thread, e.g. when
a nio thread has encounterd an exception like EOFException, through
the ErrorHandler the NioOutboundPipeline.close() method is called
from the IO thread.

If we then schedule a closetask to be scheduled on the same NIO thread
there will be a tmp self deadlock because the only thread that can
execute this task, is blocked because it is waiting on the completion
of this task. So we get a 3s stall on the IO thread; which can have
massive performance implications!

So the fix is to always offload closing to a channel to an
executor so the NIO thread will never get blocked.