Add ProbeBuilder to MetricsRegistry (#13153)
This commit adds a structure to metric name. The ProbeBuilder is used to 
register metrics with such structured name. It is backwards compatible: 
new format is enclosed in `[]`. However, the Metrics registry never 
parses the metric names, it handles them literally. The structure will 
be used by the new Jet ManCenter, and in the future by IMDG ManCenter.