Fixed ConfigXmlGeneratorTest for Windows
The ConfigXmlGenerator writes the absolute path in the XML output. The
given new File("/nonexisting-xxx") is an absolute path in Linux, but on
Windows this is changed to "C:\nonexisting-xxx", which doesn't match the
given input.

A solution for both operating systems is to create a file with a
relative path and give the absolute file from that as test input.

(cherry picked from commit a9ed3a9)