Fixed OperatingSystemMetricSetTest for Windows
The file descriptor sensors are not implemented on Windows.

The according checks for those sensors can be guarded with a test
assumption, just like the processCpuLoad and systemCpuLoad for the
UnixOperatingSystem implementation test.

I've also removed the comments that the OperatingSystemMetricSet is
hard to test. The code coverage is at 95% and it's easy to add a test
assumption whenever we see a test failing in a new test environment.

(cherry picked from commit db912e3)