Fix clientDisconnected event miss
Event firing is missed in case of a race between
ClientEngineImpl.connectionRemoved and
ClientHeartbeatMonitor.cleanupEndpointsWithDeadConnections.

Both trying to remove an endpoint(`removeEndpoint`).
If cleanupEndpointsWithDeadConnections happens to remove first,
then event is not fired.

As a fix, I moved to event firing to more centralized place so
that first call to `removeEndpoint` will fire the disconnected
event.

fixes https://github.com/hazelcast/hazelcast/issues/12674