Reset local metadata of partition upon run of clear-all operation.
Clear operation runs on every partition but only creates invalidation
on the owner partition of data-structures name. If this owner is on a
suddenly killed node and if there is a running clear operation on the
cluster at the same time, we can be in a situation that no invalidation is
generated because the owner partition of data-structures name was on
the killed node and other nodes not yet aware of the new owner for that
partition. This may cause indefinitely seen stale values on near
cache side.

With this PR, we are resetting local metadata of all partitions upon
run of clear operation so reconciliation task will manage to remove
stales.