Add ability to WAN sync only entries updated in a certain time frame (#13218)
WAN sync always transferred all entries for all partitions. Sometimes
this is too much data and the user might roughly know the time frame for
which he wants to sync entries. For instance, this may be the case when
there was an outage and the user needs to sync only entries changes
during that outage.
We now add the ability to define the from and to timestamp when syncing
entries. The existing WAN sync (without timestamps) will continue to
work in mixed version clusters but trying to sync entries in a certain
time frame will pass the initial check and fail later. In a cluster
where all members support the timestamped WAN sync feature, there should
not be any failures.