Fixed a couple of CheckStyle suppressions
Some suppressions were in suppressions.xml were superfluous or their
scope could be limited to a method. Moved them to annotation filters.

A lot of CheckStyle comment filters were not needed (CHECKSTYLE:OFF|ON).
I also disabled the comment filter, since it's not a good idea to
suppress everything. The code proves, that they are not necessary.
An annotation filter with a single or couple of suppressions is enough.