Add field for stopping WAN replication to ChangeWanStateRequest (#13275)
`stop` field is added to `ChangeWanStateRequest`. It is used to
distinguish stopping from pausing (if it's true, WAN replication is
stopped, if it's false, it's paused).

Hardcode version that's sent to MC as 3.9.4-c1, as MemberVersion
 doesn't have qualifier and adding it is out of scope for this release.

* Fix tests

Most of the failures were due to the hardcoding of version on
MemberVersion#toString.

Two tests (LocalWanPublisherStatsTest & LocalWanStatsImplTest) were
already failing on 3.9.4-c1 branch.