Dump class histogram on test failure
Introducing a JUnit rule which dumps a class histogram on test failure.
It's best-effort only as there is no public Java api to do this.

The idea is this could help to better understand test failure related
to memory utilization / GC pressure.

In theory in could be hooked into the JitterRule and dump the histogram
only when significant jitter was recorded. I decided to keep it simple
for now.