Allow HashSlotArray to work with unsigned memory addresses
Previously the sign bit was used in the ensure methods to indicate if
the value block was newly assigned or if there was an existing block for
the given key. The information if there was a new assignment or not is
localized to the methods calling ensure methods and is irrelevant
to most of the hot restart code. This allows us to encode this
information differently. The ensure methods now return a flyweight
container object with both the unsigned address and a boolean indicating
if there was a new assignment.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2175