Add RestartableException & fix HazelcastWriters (#907)
We ignored the `HazelcastInstanceNotActiveException` in
`HazelcastWriters`, if the object written to was local. We relied that
the job will restart anyway and writing will work after that. But if the
job completed after such failure, but before a restart, the writing
error was silently ignored.

This commit fixes the problem by introducing `RestartableException` that
can be thrown from processors to indicate a failure that can be resolved
by a restart.