Fixes #13333
In the original test, by the time `setTTL` operation runs some of the entries are already expired due to short initial TTL.
This PR increases the initial TTL and decreases the number of entries. We still need to wait initial TTL amount of time to make sure the entries are not evicted at the end of TTL duration.