Fix exceptions thrown to the user
Retrieving a service instance while shutting down the node throws a
RetryableHazelcastException. This may make sense for most cases like
running an operation as the operation can then be retried. In some cases
like creating a proxy, this exception is thrown directly to user code.
In this case, the code is not retriable as the node will shut down and
will never be able to create a proxy.

We fix this issue by catching the exception thrown when retrieving the
service instance and rethrowing it as
HazelcastInstanceNotActiveException. This is what most of the proxy code
throws in case the instance is not active.

The alternative fix which might have been more suitable would have been
for the getService call to throw a HazelcastInstanceNotActiveException
and for the callers to rethrow a retryable exception in case a retry was
needed. Unfortunately, we have several issues with this solution:
- NodeEngine.getService is part of SPI and the javadoc states that it
throws a HazelcastException if a service is not available
- HazelcastInstanceNotActiveException does not extend HazelcastException
and we cannot make it extend it since it extends IllegalStateException
already
- changing the exception thrown from the getService method would require
extensive changes to the code calling this method (checking each call
and determining if we can retry). This is not an issue in itself but it
should have been done from the beginning. Now we cannot change the
exception because of the above two points.

Also, changed the parent class for InvalidConfigurationException to
HazelcastException as we advertise HazelcastException to be a base class
for hazelcast exceptions. The InvalidConfigurationException still
extends RuntimeException so this is acceptable.

Fixes: https://github.com/hazelcast/hazelcast/issues/13225