Fix handling of local Node ID overflow [backport of #13330] (#13331)
If local Node ID was overflowed, the `FlakeIdGeneratorProxy.getNodeId`
was supposed to return -2 (`NODE_ID_OUT_OF_RANGE`). There is an
optimization in it which caches the nodeId value. However, it cached the
overflowed nodeId, not -2. So on subsequent call it returned overflowed
id, on which the `newIdBaseLocal` failed with `AssertionError: nodeId
out of range: NNN`.

The fix is to assign correct value to `FlakeIdGeneratorProxy.nodeId`.

The test supposed to detect this didn't due to a bug that the non-smart
client always chose port 5001 (the `instance1`) and never went to
`instance2`. So the false success probability of 0.1% was actually 100%.