Provide a way to force service name serialization
Consider an operation that doesn't override its getServiceName, i.e the
service name is provided by an explicit external call to setServiceName.
The operation is already released as a part of an official HZ release.
If, in the next version, one decided to override getServiceName to
provide the same service name, the serialized representation of the
operation is no longer compatible with the previous version since the
service name is not serialized if it matches the one returned by
getServiceName.

This change provides a way to work around this issue and fixes several
affected operations.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2198