Provide WAN replication publisher state through JMX
We provide the WAN replication publisher state through JMX. The managed
bean is somewhat different from the existing mbeans:
- there can be many WAN publisher mbeans (unlike mbeans for some
services)
- a single WAN publisher is identified with a composite identifier - WAN
 replication name + target cluster group name. This is different from
other mbeans with a single type (e.g. lock mbeans) which have only one
identifier

An another complication is that the OS and EE WAN publisher don't have a
 common interface. To avoid introducing such and interface and introduce
larger refactoring, we will expose the WAN publisher mbean in such a
way as if there was a common interface and internally hide the
complexity. Later on, we can clean it up and introduce the common WAN
publisher interface for OS and EE. The only issue is that the MBean
managed object class name may change at that point.