Refactored some CheckStyle suppressions to annotation based style (#13377)
Some classes already had mixed suppression styles, e.g. Bits. Other
suppressions like NPathComplexity should not be applied on a whole
class.

In general it makes more sense to see the CheckStyle suppressions
directly in the code, so they can be considered when refactoring a
class. The suppressions.xml is easily forgotten when changing code.