Fixed TransactionalMapProxy with Near Cache
The Near Cache key for a TransactionalMapProxy was not deserialized when
it was in binary format, which happens when a Hazelcast client is used.

Another problem was that the Near Cache key was created first, even if
it was not used at all. Especially with this fix, we might run into an
unused deserialization, when no Near Cache was configured.

The order of read-only calls is:
* read value from internal TXN map (always binary key)
* read value from Near Cache (binary/object key)
* read value from remote (always binary key)

The Near Cache key creation has been moved behind the TXN map lookup.
This prevents the key creation if it's not used at all. It uses the
original and binary key, to prevent a duplicated (de)serialization.

The order of write/remove calls is:
* remove the value (always binary key)
* invalidate the Near Cache (binary/object key)

The Near Cache key creation has been moved to the invalidation method,
so behind all checks if the Near Cache is enabled. This prevents the
key creation if it's not used at all.