During cluster-state change, trigger repartitioning only when needed
Currently, repartitioning is always triggered during cluster change,
regardless of whether or not it's needed. But since repartitioning algorithm
is not deterministic, it can create one or two migrations on every call.

This can cause redundant migrations and failures during Hot Restart validation.

Instead now, we set a flag when a repartitioning action (member add or remove) occurs,
and trigger repartitioning if that flag is set on cluster state change event.