Cleanup of Node field access
The direct access to Node fields is a problem for compatibility tests,
which used proxied or mocked classes from another classloader. Method
calls can be proxied or intercepted easily, access to fields is a real
issue.

This PR fixes the most important Node access locations to enable the
usage of Hazelcast instances in compatibility tests.