Preparing Merkle tree IMap integration in OS
The Merkle tree integration is done at `RecordStore` level via an observer, which captures all mutating interactions with the `RecordStore`.
The goal with this abstraction is to keep the RecordStore and the Merkle tree uncoupled.
The actual integration is done in EE.

This change also moves writing to the `EventJournal` behind the `RecordStoreMutationObserver`