LazyJsonValue tests and not-thread safe warning
LazyJsonValue class is not thread safe. I added a line stating this at the javadocs of Json related public interfaces.

I added LazyJsonValue equals and hashcode tests & implementation. LazyJsonValue is expected to be equal to any JsonValue that represents the same object. Also hashcodes of such equal objects are the same.