Cancel pending invocations & wait-set entries when client disconnects
When client wants to execute an operation, a member submits an invocation
on behalf of that client. And when response is received on member-side,
invocation is deregistered and response is forwarded to the client.

If that's a blocking operation, such as lock or reliable-topic message listener,
that operation is parked in a wait-set until it's notified by a notifier operation.
When client is disconnected, entry is wait-set is marked as invalid and removed
from the wait-set eventually.
But invocation on member-side is not notified
and does not receive any response, hence stays in the registry either until timeout
or forever depending on the invocation's target (local or remove)
and/or operation's type (backup-aware or not).

Instead, with this change, when a client is disconnected, wait-set entry will be cancelled
with an exception which then will be propagated to the submitting invocation as a response.
This way, pending invocation will be deregistered.

This is different from the member is removal case. Because members do not submit invocations
on behalf of other members, every member submits and keeps track of its own invocations.