Detect leaking threads in tests
This change detects leaking threads in tests' `@AfterClass`. If a leak is detected the name of the thread is reported in an `IllegalStateException` along with the leaking thread's stack trace, breaking the test. The actual test case causing the leak cannot be identified directly, but it can be found by knowing the thread's name.