Add Merkle tree Management Center endpoint
The actual implementation of the consistency check will be done in EE.
This is only the endpoint that the MC can invoke.

The consistency check can be triggered by `WanCheckConsistencyRequest`.
The result of the check will be written eventually into the `LocalWanPublisherStats#lastConsistencyCheckResults`, which is populated periodically in the `TimedMemberState`.