Provide index statistics (#13359)
Public API part includes LocalMapStats extended with the information
about the number of the queries executed on the map (getQueryCount and
getIndexedQueryCount) and with the per-index statistics (getIndexStats)
which is provided in a form of LocalIndexStats. The statistics is also
published to the metrics subsystem.

The most tricky implementation detail is the proper counting of the
queries on a per-index basis. Since a single query may hit a certain
index multiple times, a special query-detecting wrapper around the
actual index is returned by a query context. For instance, see
GlobalQueryContextWithStats.

To minimize the overhead induced by the statistics collection, almost
all implementations of the related interfaces are present in three
flavors: specialized for global indexes (which are shared among
partitions), specialized for partitioned indexes and specialized for the
statistics-off configuration.