Fix race in AsyncMapWriter (#991)
The writer uses `PartitionIteratingOperation`, which can partly fail:
some partitions succeed, some fail. The failed partitions are retried.
The retrial is executed from an `ExecutionCallback`, which is called
once for each member. Both attempts use `invokeOnCluster`, which used a
`doneLatch` to track callbacks from members and when it got to 0, it
completed the future. The problem was if in the first attempt, two
members failed. It will cause two retry calls to `invokeOnCluster`. But
the first completed retry call completed the future before the second
one finishes, causing the test to (correctly) fail.

The solution is to use one `doneLatch` for both the initial attempt and
all retries. `invokeOnCluster` only increments and decrements it. If a
partition is retried, it is first incremented by the number of retry
operations so that it gets to 0 only when all retrials are successfully
done.

The test failed in about 0.1% of runs.

Fixes #975