ChangeLoggingRule should be able to cope with all kinds of LoggerFactory
The logger factory used while executing the test suite can change
depending on system properties, maven profiles active etc. The
ChangeLoggingRule should not fail when an unexpected LoggerFactory
is instantiated; instead it will output a short message to stdout
that it was not able to change logging configuration, so the information
is not completely lost.