Increases wait time in cache expiration backup test
The test does put-remove and ensures that backups also remove the entries. It ensures that remove operation is responsible for removals but not the expirations by waiting allowing expiration amount of time to pass. When the test experiences hiccups, one second wait may still cause expirations. So this pr increases the expiration time to 5 seconds to allow for hiccups.