Fix race when collecting responses from members (#1030)
The `future` from the `invoke` method was added to the map at a time the
callback could already be executed. This will cause that some of the
responses will be missing when handling.

Fixes #1029: from the logs we see that only one of the member failed and
the error response from the failing member was indeed sent, but the
terminal snapshot was marked as successful despite of that. This
suggests that the response was lost, even though it never reproduced
locally. The window for race is quite small: the thread, that should
execute as the second one, has lot more work to do than the other one.