Fix ClientDisconnectTest failure
In the failing tests, the client is issuing an async operation and then terminating afterwards. If the client termination occurs before the invocation is started & registered on the server, that invocation will not be notified during handling termination of the client. Therefore, before terminating the client, tests should wait until its operations are registered on the server.

Fixes #13680