Throw HazelcastInstanceNotActive when serialization service is disposed
Fixes #13724

Original problem:
When serialization service is disposed then DataSerializableSerializer
and PortableSerializer clear their internal factories. Hence if there
is another thread asking to deserialize Data then a factory won't be found
and it will throw SerializationException. This Exception is misleading
as in it indicates there is something wrong with Serialization config
while in the fact it's just HazelcastInstance shutting down.