Dont keep the endpoint reference to clean client resources
We are scheduling a task to clean client resources after some
time(60 seconds). Eventough endpoint is removed, we were keeping
a reference to it via this task. This was causing unncessary leak
for 60 seconds since the endpoint keeps reference to heavy objects
like `connection`.

With this pr, we are passing only necessary fields objects of
client endpoint to task to avoid keeping reference to the client
endpoint itself.