Add WAN publisher group name property (#13721)
Until now, the "group-name" attribute was used both as an identifier for
a specific publisher in a WAN replication scheme and as a group name
for authentication. This does not allow replicating to two distinct
clusters with the same group name as the latter one will be overwritten.

We add the ability to configure the group name for authentication in the
publisher group properties. The "group-name" attribute still must be
unique but now it can be used only for identification of a single
publisher in a WAN replication scheme.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2393