Introduce WAN publisher ID
Introduces a WAN publisher ID which is used to refer to a specific WAN
publisher in a WAN replication scheme. If a WAN publisher ID is not
used, the group name is used as a publisher ID.
The WAN publisher ID may be useful when the target group names are not
unique for all of the WAN replication publishers in a single WAN
replication scheme.

Fixes:
https://github.com/hazelcast/hazelcast-enterprise/issues/2419
https://github.com/hazelcast/hazelcast-enterprise/issues/2423