Removes stale caching of ringbuffer in AbstractRingBufferOperation
AbstractRingBufferOperation caches the ringbuffer that it operates on. This may cause operations running on wrong ringbuffer container because RingBuffer.ReadMany operation is a long living operation and partition migrations may cause cached ringbuffer container to go stale. When the partition that the ringbuffer is located in migrates, reference to ringbuffer container becomes stale. However, it does not create an adverse effect because operation service detects that the required partition is not in the same instance that ReadMany is called from anymore. So the operation service does a remote operation. However, later the same partition may come back to the initial instance in the future due to further changes in network. This time ringbuffer operations are local and they use the stale ringbuffer container.

This PR removes unnecceassary caching that causes stale container.