Wait for safe state before destroying cache (#13805)
Cache destroy and migrations are inherently racy, as for all of the
structures. In the test, after a second instance is created, some
partitions are replicated to the new member. This member is already
in the process of destroying the cache. The migration can then
re-create the record store after the destroy mechanism already checks
if the migrating partition contains the record store to be removed.

Since all of the data structures suffer from this problem, it should be
addressed systematically. We then fix the test to wait for migrations to
complete before proceeding to destroy the cache.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/930