Fix possible NullPointerException in AbstractCacheRecordStore#canSendBackupExpiration
A cache entry is removed at `get` time if it is found to be expired. The member sends backups expiration events to evict the backup entries for it too. If the backup member leaves the cluster between the detection and creation of eviction operation, AbstractCacheRecordStore#canSendBackupExpiration may find related member object to be null. This causes NullPointerException and the expiration event is not sent. Moreover HiDensityNativeMemoryCacheRecordStore's onProcessExpiredEntry method is not called. This leads to corresponding entry to be removed from the record store but its memory is not deallocated. This leads to memory leak.