Increase TTL and add assertion to clarify failure cause for racy test
The test adds entries and tries to remove them immediately afterwards
before TTL has expired. If there is a big pause between adding and
removing entries, they may expire before they are removed. We increase
the TTL and add an assertion message to clarify the cause.

If the problem persists, we may have to change the test or remove
altogether.

Fixes: https://github.com/hazelcast/hazelcast/issues/13619