Fixes blocking RingBuffer ReadManyOperation
ReadManyOperation(async) blocks when there is no item in the ringbuffer. Then the thread is parked. When an add operation adds new items, it attempts to unblock waiting read operations. Before unparking the op, the system calls the op's shouldWait method again. shouldWait throws because now sequence is stale (AddAll operation added too many items and overwritten the first one). Since it throws during unparking, the exception is not sent to the requesting party. It is logged and dropped instead. With this PR, shouldWait does not throw anymore. Instead, it lets the operation continue to throw in beforeRun method. Therefore all exceptions are sent to the calling party.