Skip expired events in journal when reading (#13749)
Skip expired events in journal when reading

The event journal read operation would clean up expired events in two
places - before executing the operation and each time events would be
added to the result set. As the read operation is loss-tolerant, the
requested sequence was corrected only after the first cleanup. Since
there is a time window between these two cleanups, the sequence needs to
be readjusted after the second cleanup as well.
This might also mean that the result set may contain "gaps" if there
were big pauses and the items expired after they have been added to the
journal and before they were added to the result set.

Fixes: https://github.com/hazelcast/hazelcast/issues/13746