Fixes inconsistent config in EntryProcessorBouncingNodesTest
Bouncing nodes are started with indexes on in withPredicateNoIndex test. Inconsistent config in cluster causes entry processors to believe there are indexes whereas there aren't any. So some entries cannot be found even though they are present in the record stores.

EntryProcessorBouncingNodesTest is also refactored. The test scenario is exactly the same but assertions are simpler.