Fix terminal snapshot when one is in progress
The JobCoordinationService called `shouldStartJobs` to check whether to
start snapshots. It contained the condition that it doesn't start when
member is shutting down or migration is ongoing. These two conditions do
not apply to snapshots, however. For example, when a terminal snapshot
is started and another snapshot was ongoing, it was postponed. The next
snapshot was supposed to be terminal, but it couldn't start, because
member is shutting down...