Retry map loading when a map load result completes a subsequent load (#13890)
The map.get operation will trigger an asynchronous load process where
the result of the load process is completed for some partitions twice
(as seen in MapKeyLoader.sendKeyLoadCompleted).
A subsequent load can then observe the first completion and start
loading. At this point, the subsequent load can get completed from a
result of the previous load process. This is why a map.size operation
might throw an exception from the first load process. In this case, we
log a message and retry loading until the map shows the expected size.

Some alternatives could be:
- waiting until all partitions have been notified of the result of the
first load process before initiating a second load process.
Unfortunately, we can't detect this. As some partitions are completed
twice, we might just observe the first completion and run into the same
issue again.
- add some waiting after map.get and before the subsequent load. This is
a brittle fix as any waiting may be thwarted by a large enough pause
- return from map.get once all partitions have been notified. This is a
substantial change and could severely affect the performance of the get
operation since now it has to wait for all partitions to acknowledge the
load result before proceeding

Fixes: https://github.com/hazelcast/hazelcast/issues/5453