Ignore/fix broken index stats selectivity tests
Consider the following setup: a map with 10 entries, like 1 -> 1, 2 -> 2,
..., which is hosted by 2 members and the partition count is much higher
than 10. We are really unlucky and the first member is hosting only a
single map entry, say 1 -> 1; all other entries are hosted by the second
member.

Now, if we query for the 1 -> 1 entry using indexes, on the first member
the selectivity will be 0.0 (we are selecting "all" entries) while on
the second member it will be 1.0 (we are selecting nothing), but the
expected selectivity is 0.9.

No matter what we do, we can't recover the expected selectivity without
going into much details about partitions and entries distribution among
cluster members for each query in question.

Fixes: https://github.com/hazelcast/hazelcast/issues/13657