Made entry removal and stats update atomic.
More than one method can attempt to change a stats counter
concurrently. For example, for owned entry counter, it is possible that
near-cache clear() method sets it to 0, then near-cache remove() decrements it.
In the cases like this, we can see a negative value in stats. To prevent this,
we are making entry-removal and stats-update atomic.