Fixes OOME due to accapted connections
Fixes
https://github.com/hazelcast/hazelcast-enterprise/issues/2492
The cause of the problem is the new pipeline in 3.11.
In the old approach, the channel wasn't registered before the
connection was established.
So in case of failure, nothing needs to be unregistered.
But with the new pipeline the channel gets registered (created)
before the connection is established
but it didn't get unregistered of the connection could not be
established. Leading to a memory leak.