Fixes LazySamplingEntryIterableIterator
This iterator starts iterating from a random segment and a random bucket in that segment. When the only item is located in the starting point, the iterator fails to progress after returning the first element. It keeps resetting `currentEntry` to the first and only element in the bucket at each iteration. That is because `currentEntry` has 3 different meanings depending on the state of the iteration:
1. We havent checked any item in the current bucket (currentEntry is null)
2. We exhausted all the items in the current bucket (currentEntry is null)
3. There is something to return from this bucket ( currentEntry is not null)

We cannot determine if we are at state 1 or 2. The code assumes that we are at state 1 when in doubt.

Now, currentEntry always points to the latest returned entry in the current bucket, so it can mean only 2 things
1. We havent checked any item in the current bucket ( currentEntry is null )
2. We returned an item from this bucket ( currentEntry is not null)